+++ libs/charts/base/src/lib/components/column/model.ts
@@ -1,0 +1,5 @@
+import { Column, ColumnOptions } from '@antv/g2plot';
+
+export type PrizmChartsColumnOptions = ColumnOptions;
+export type PrizmChartsColumnOrigin = Column;
+export type PrizmChartsColumnItem = Record<string, unknown>;